feat(README): update instructions and remove pytorch example
This commit updates the instructions for using Singularity containers.
It removes the specific example of creating a sandbox container from
the Pytorch Docker image. It also adds a new section "1.5 Try it out"
to encourage users to experiment with customizing container images.
Lastly, it corrects a command syntax in the section about loading
.bashrc inside the container.